- BUG: Data summaries were wrong when the number of alleles broke the 32bit INT limit.

+++ trunk/src/jhi/flapjack/analysis/AlleleStatisticsTest.java
@@ -45,10 +45,10 @@ public class AlleleStatisticsTest extends TestCase
 		AlleleStatistics statistics = new AlleleStatistics(viewSet);
 		statistics.runJob(0);
 
-		ArrayList<int[]> results = statistics.getResults();
+		ArrayList<long[]> results = statistics.getResults();
 
 		// This viewset only has one view
-		int[] array = results.get(0);
+		long[] array = results.get(0);
 
 
 		assertEquals(array[0], 0);	// # of unknowns
